--- src/main/java/com/admios/connector/watsonvisualrecognition/exceptions/VisualRecognitionException.java
@@ -1,15 +1,0 @@
-package com.admios.connector.watsonvisualrecognition.exceptions;
-
-public class VisualRecognitionException extends Exception {
-
-	private static final long serialVersionUID = 1L;
-
-	public VisualRecognitionException() {
-		super();
-	}
-	
-	public VisualRecognitionException(String message){
-		super(message);
-	}
-
-}
