--- src/main/java/com/admios/connector/watsonvisualrecognition/handler/CommonHandler.java
@@ -1,14 +1,0 @@
-package com.admios.connector.watsonvisualrecognition.handler;
-
-import com.ibm.watson.developer_cloud.visual_recognition.v3.VisualRecognition;
-
-public abstract class CommonHandler<T> {
-
-	protected final VisualRecognition service;
-
-	public CommonHandler(VisualRecognition service) {
-		this.service = service;
-	}
-
-	abstract public T execute();
-}
