--- src/main/java/com/admios/connector/watsonvisualrecognition/handler/implementation/RetrieveClassifierDetailsHandler.java
@@ -1,20 +1,0 @@
-package com.admios.connector.watsonvisualrecognition.handler.implementation;
-
-import com.admios.connector.watsonvisualrecognition.handler.CommonHandler;
-import com.ibm.watson.developer_cloud.visual_recognition.v3.VisualRecognition;
-import com.ibm.watson.developer_cloud.visual_recognition.v3.model.VisualClassifier;
-
-public class RetrieveClassifierDetailsHandler extends CommonHandler<VisualClassifier> {
-
-	protected String classifierId;
-
-	public RetrieveClassifierDetailsHandler(VisualRecognition service, String classifierId) {
-		super(service);
-		this.classifierId = classifierId;
-	}
-
-	@Override
-	public VisualClassifier execute() {
-		return service.getClassifier(classifierId).execute();
-	}
-}
