+++ src/main/java/org/mule/modules/watsonvisualrecognition/exceptions/VisualRecognitionException.java
@@ -1,0 +1,15 @@
+package org.mule.modules.watsonvisualrecognition.exceptions;
+
+public class VisualRecognitionException extends Exception {
+
+	private static final long serialVersionUID = 1L;
+
+	public VisualRecognitionException() {
+		super();
+	}
+	
+	public VisualRecognitionException(String message){
+		super(message);
+	}
+
+}
