--- src/test/java/org/mule/modules/watsonalchemylanguage/automation/runner/SystemTestSuite.java
@@ -1,12 +1,0 @@
-package org.mule.modules.watsonalchemylanguage.automation.runner;
-
-import org.junit.runner.RunWith;
-import org.junit.runners.Suite;
-import org.junit.runners.Suite.SuiteClasses;
-import org.mule.modules.watsonalchemylanguage.automation.system.ConnectorConfigTestCases;
-
-@RunWith(Suite.class)
-@SuiteClasses({ ConnectorConfigTestCases.class })
-public class SystemTestSuite {
-
-}
