+++ src/test/java/org/mule/modules/watsonvisualrecognition/automation/runner/SystemTestSuite.java
@@ -1,0 +1,12 @@
+package org.mule.modules.watsonvisualrecognition.automation.runner;
+
+import org.junit.runner.RunWith;
+import org.junit.runners.Suite;
+import org.junit.runners.Suite.SuiteClasses;
+import org.mule.modules.watsonvisualrecognition.automation.system.ConnectorConfigTestCases;
+
+@RunWith(Suite.class)
+@SuiteClasses({ ConnectorConfigTestCases.class })
+public class SystemTestSuite {
+
+}
